Allow data to be saved for the original post being edited Closes #26
This is caused by the wonderful complexity of action hooks:

- ACF adds a hook onto the `save_post` action in order to update all ACF data. Makes sense.
- Stream Manager adds a hook onto `transition_post_status`, which is fired *before* `save_post` is.
- Stream Manager's callback, `on_save_post` calls the Stream's `insert_post` method, which happens to make a call to `wp_update_post`, whiiich happens to fire `save_post`. *Whew*
- This causes ACF to pick up on the `save_post` hook that is then fired (currently for the stream, *not* the original post being edited.
- During ACF's `save_post` callback(s), it checks for the presence of an ACF generated nonce. If it finds it, it actually modifies the `$_POST` array and sets it to false.
- When `save_post` is fired for the post actually being edited, and not the stream, the `$_POST `array no longer contains a valid nonce, so ACF bails.

This PR is a bit of a hacky way to get around that - in Stream Manager's `transition_post_status` method, it takes the liberty of firing `save_post` *before* any calls are made to `wp_update_post` so that other actions have a chance to work with the original post being edited. It then removes all currently set actions for the `save_post` callback so they aren't fired twice.