Add visual limit to number of stories in Stream interface Towards #28 Since objects are removed from the dom when the "x" is clicked, a simple css selector can encompass the visual requirements of this enhancement.
What is *not* covered is the case when the number of posts left in the entire stream are less than or equal to the `stream-manager_stub_display_limit` *before* the "x" is clicked. If that happens, there is no post to take it's place. I assume this is intended behavior, but if the intention is to have the empty spot automatically/immediately populated by querying for a new post, let me know @jarednova and I'll tackle that.